\docType{data}
\name{food_composition_data}
\alias{food_composition_data}
\title{Nutritional values for common foods and products}
\format{List}
\description{
This list contains 3 different food composition tables, which provide information 
on the average nutritional value of foods consumed in United States (USDA standard
reference database), France (CIQUAL database) and Spain (BEDCA database). All nutrition
information is provided per 100 grams of food. 
}
\references{
https://ndb.nal.usda.gov/ndb/

http://www.bedca.net/

https://ciqual.anses.fr/
}
\usage{data(food_composition_data)}
\value{List}
