\docType{data}
\name{NIH_nutrient_recommendations}
\alias{NIH_nutrient_recommendations}
\title{Nutrient recommendations}
\format{List}
\description{
This list contains nutrient recommendations from the NIH (National Institutes of Health)
database:

- The first element of the list contains the recommended dietary allowance (RDA)
of 33 nutrients, by gender and age.

- The second element of the list contains the tolerable upper intake level (TUIL)
of 30 nutrients, by gender and age.
}
\usage{data(NIH_nutrient_recommendations)}
\value{List}
