\docType{data}
\name{nutrient_group}
\alias{nutrient_group}
\title{Nutrient groups}
\format{Matrix}
\description{
This matrix contains nutrient names and groups, for all nutrients included in
the NIH_nutrient_recommendations dataset.
}
\references{
https://www.nih.gov/
}
\usage{data(nutrient_group)}
\value{List}
