\docType{data}
\name{sample_diet_USDA}
\alias{sample_diet_USDA}
\title{Example of a one-week diet}
\format{List}
\description{
This list is an example of a one-week diet, using foods from the USDA database.
Each element of the list is a matrix, which includes the all the foods eaten in
a given day.
}
\references{
https://ndb.nal.usda.gov/ndb/
}
\usage{data(sample_diet_USDA)}
\value{List}
